configs: qemu-gdb: disable CONFIG_RELOCATABLE
Since RELOCATABLE is set to default 'y' now. It has to be disabled only
due to that we run out of memory since CONFIG_RELOCATABLE adds some
bloat to the kernel Image, so if we build are really big kernel (like
building an allmodconfig kernel) it might look like CONFIG_RELOCATABLE
is broken, but all it does is pushing it over the edge.

Fixes: dd4bc6076587 ("arm64: warn on incorrect placement of the kernel by the bootloader")
Signed-off-by: Anders Roxell <anders.roxell@linaro.org>