tools/build: Add clang cross-compilation flags to feature detection
When a tool cross-build has LLVM=1 set, the clang cross-compilation
flags are not passed to the feature detection build system. This
results in the host's features are detected instead of the targets.

E.g, triggering a cross-build of bpftool:

  cd tools/bpf/bpftool
  make ARCH=riscv CROSS_COMPILE=riscv64-linux-gnu- LLVM=1

would report the host's, and not the target's features.

Correct the issue by passing the CLANG_CROSS_FLAGS variable to the
feature detection makefile.

Fixes: cebdb7374577 ("tools: Help cross-building with clang")
Signed-off-by: Björn Töpel <bjorn@rivosinc.com>
Acked-by: Jiri Olsa <jolsa@kernel.org>